main.js 모듈 분리 리팩토링 (#1)
2384줄 단일 파일을 논리적 단위 9개 모듈로 분리:
- state.js: 게임 상태 및 캔버스 참조
- utils.js: 유틸리티 함수 (그리드, 색상, 수학, 포탑 계산)
- map.js: 경로 데이터 및 건설 가능 여부 판정
- audio.js: Web Audio 엔진
- ui.js: DOM 참조 및 UI 관리 함수
- game.js: 게임 로직 및 업데이트 루프
- renderer.js: 렌더링 파이프라인
- main.js: 이벤트 바인딩 및 게임 루프 (엔트리포인트)

기존 constants.js, data/enemyStyles.js, data/towerTypes.js 유지.
ES Module import/export 기반, 스모크 테스트 통과.

Closes #1
