+++ package.json
@@ -1,20 +1,19 @@
-﻿{
-    "name":  "tower-defense",
-    "version":  "1.0.0",
-    "main":  "main.js",
-    "directories":  {
-                        "doc":  "docs"
-                    },
-    "scripts":  {
-                    "test":  "node tests/smoke.test.js"
-                },
-    "keywords":  [
-
-                 ],
-    "author":  "",
-    "license":  "ISC",
-    "description":  "",
-    "dependencies":  {
-                         "jsdom":  "^27.0.0"
-                     }
+{
+    "name": "tower-defense",
+    "version": "1.0.0",
+    "main": "src/main.js",
+    "directories": {
+        "doc": "docs"
+    },
+    "scripts": {
+        "test": "node tests/smoke.test.js"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "description": "",
+    "dependencies": {
+        "jsdom": "^27.0.0"
+    },
+    "type": "module"
 }
