+++ src/data/towerTypes.js
@@ -1,0 +1,394 @@
+﻿export const TOWER_TYPES =  {
+    basic: {
+        id: "basic",
+        label: "기본",
+        attackPattern: "projectile",
+        cost: 35,
+        baseDamage: 20,
+        baseUpgradeCost: 40,
+        range: 165,
+        fireDelay: 0.6,
+        projectileSpeed: 520,
+        projectileLife: 0.85,
+        projectileRadius: 6,
+        trailLength: 60,
+        shape: "turret",
+        baseColor: "#1c2740",
+        coreColor: "#dae8ff",
+        barrelColor: "#9abaf9",
+        glowColor: "#8ab7ff",
+        auraColor: "rgba(138, 183, 255, 0.32)",
+        muzzleFlashColor: "#fde7a2",
+        muzzleLengthMultiplier: 1.7,
+        flashSizeMultiplier: 0.95,
+        recoilKick: 0.5,
+        recoilRecovery: 3.1,
+        turnSpeed: 7.2,
+        outline: "#1f2f5a",
+        levelColors: [
+            "#bdd6ff",
+            "#a2c0ff",
+            "#8aa9ff",
+            "#7294ff",
+            "#5a7eff",
+            "#4368ff",
+            "#2e56f2"
+        ],
+        projectileColors: [
+            "#f8faff",
+            "#e5f0ff",
+            "#cddfff",
+            "#b3cdff",
+            "#9abbff",
+            "#81a7ff",
+            "#6b95ff"
+        ]
+    },
+    shotgun: {
+        id: "shotgun",
+        label: "샷건",
+        attackPattern: "shotgun",
+        cost: 55,
+        baseDamage: 6,
+        baseUpgradeCost: 45,
+        range: 150,
+        fireDelay: 0.9,
+        pellets: 5,
+        spread: Math.PI / 3.5,
+        projectileSpeed: 540,
+        projectileLife: 0.55,
+        projectileRadius: 6,
+        trailLength: 40,
+        shape: "vulcan",
+        baseColor: "#0c233d",
+        coreColor: "#ffe3a3",
+        barrelColor: "#f5a844",
+        glowColor: "#ff985c",
+        auraColor: "rgba(255, 152, 92, 0.32)",
+        muzzleFlashColor: "#ffd17a",
+        muzzleLengthMultiplier: 1.45,
+        flashSizeMultiplier: 1.05,
+        recoilKick: 0.62,
+        recoilRecovery: 2.9,
+        turnSpeed: 6.4,
+        outline: "#0b2b4c",
+        levelColors: [
+            "#90caf9",
+            "#74b9f6",
+            "#4fa8f4",
+            "#2d96f1",
+            "#1f86e8",
+            "#136fd6",
+            "#0d5abf"
+        ],
+        projectileColors: [
+            "#ffe082",
+            "#ffd54f",
+            "#ffca28",
+            "#ffc107",
+            "#ffb300",
+            "#ffa000",
+            "#ff8f00"
+        ]
+    },
+    longrange: {
+        id: "longrange",
+        label: "저격",
+        attackPattern: "beam",
+        cost: 90,
+        baseDamage: 35,
+        baseUpgradeCost: 65,
+        range: 260,
+        fireDelay: 1.6,
+        projectileSpeed: 820,
+        projectileLife: 1.5,
+        projectileRadius: 6,
+        trailLength: 150,
+        beamWidth: 6,
+        shape: "rail",
+        baseColor: "#1a1034",
+        coreColor: "#cebaff",
+        barrelColor: "#8d72ff",
+        glowColor: "#a57eff",
+        auraColor: "rgba(165, 126, 255, 0.28)",
+        muzzleFlashColor: "#e4d6ff",
+        muzzleLengthMultiplier: 2.2,
+        flashSizeMultiplier: 0.8,
+        recoilKick: 0.35,
+        recoilRecovery: 3.4,
+        turnSpeed: 8.6,
+        outline: "#27144a",
+        levelColors: [
+            "#d0bbff",
+            "#ba9dff",
+            "#a279ff",
+            "#8c5aff",
+            "#7b43f0",
+            "#6a31d3",
+            "#5925b8"
+        ],
+        projectileColors: [
+            "#f5ebff",
+            "#e8d4ff",
+            "#dcbcff",
+            "#cea4ff",
+            "#c08cff",
+            "#b073ff",
+            "#a05dff"
+        ]
+    },
+    burst: {
+        id: "burst",
+        label: "점사",
+        attackPattern: "burst",
+        cost: 65,
+        baseDamage: 14,
+        baseUpgradeCost: 55,
+        range: 175,
+        fireDelay: 1.05,
+        burstCount: 3,
+        burstDelay: 0.08,
+        projectileSpeed: 600,
+        projectileLife: 0.85,
+        projectileRadius: 7,
+        trailLength: 70,
+        shape: "prism",
+        baseColor: "#341d0a",
+        coreColor: "#ffbe73",
+        barrelColor: "#ff8e3f",
+        glowColor: "#ff9f52",
+        auraColor: "rgba(255, 159, 82, 0.28)",
+        muzzleFlashColor: "#ffd9a1",
+        muzzleLengthMultiplier: 1.6,
+        flashSizeMultiplier: 0.9,
+        recoilKick: 0.48,
+        recoilRecovery: 3,
+        turnSpeed: 7.8,
+        outline: "#4a2509",
+        levelColors: [
+            "#ffd180",
+            "#ffb74d",
+            "#ffa040",
+            "#ff8f2c",
+            "#ff7b12",
+            "#f56705",
+            "#e25900"
+        ],
+        projectileColors: [
+            "#fff3cd",
+            "#ffe7a3",
+            "#ffd26e",
+            "#ffbe42",
+            "#ffaa23",
+            "#ff9514",
+            "#ff8609"
+        ]
+    },
+    rapid: {
+        id: "rapid",
+        label: "연사",
+        attackPattern: "projectile",
+        cost: 45,
+        baseDamage: 10,
+        baseUpgradeCost: 45,
+        range: 145,
+        fireDelay: 0.22,
+        projectileSpeed: 640,
+        projectileLife: 0.65,
+        projectileRadius: 5,
+        trailLength: 75,
+        shape: "gyro",
+        baseColor: "#052b18",
+        coreColor: "#8ce7b5",
+        barrelColor: "#43d886",
+        glowColor: "#66f3b6",
+        auraColor: "rgba(82, 235, 164, 0.32)",
+        muzzleFlashColor: "#bbffd9",
+        muzzleLengthMultiplier: 1.55,
+        flashSizeMultiplier: 0.85,
+        recoilKick: 0.38,
+        recoilRecovery: 3.6,
+        turnSpeed: 9.2,
+        outline: "#073519",
+        levelColors: [
+            "#b9f6ca",
+            "#9bf1b3",
+            "#78e99a",
+            "#55dd80",
+            "#3ccc6b",
+            "#2daf58",
+            "#1f9348"
+        ],
+        projectileColors: [
+            "#d0ffdc",
+            "#aefecc",
+            "#88fcb5",
+            "#5af79a",
+            "#2fef7f",
+            "#1ddc6d",
+            "#10c05b"
+        ]
+    },
+    explosive: {
+        id: "explosive",
+        label: "폭발",
+        attackPattern: "explosive",
+        cost: 80,
+        baseDamage: 26,
+        baseUpgradeCost: 70,
+        range: 170,
+        fireDelay: 1.25,
+        projectileSpeed: 480,
+        projectileLife: 1.1,
+        projectileRadius: 8,
+        explosionRadius: 85,
+        explosionColor: "rgba(255, 182, 89, 0.45)",
+        explosionHaloColor: "#ffe9c0",
+        explosionStrokeColor: "rgba(255, 161, 86, 0.55)",
+        explosionLife: 0.5,
+        trailLength: 55,
+        shape: "howitzer",
+        baseColor: "#3e1b08",
+        coreColor: "#ffbc76",
+        barrelColor: "#ff8a45",
+        glowColor: "#ffa25a",
+        auraColor: "rgba(255, 162, 90, 0.32)",
+        muzzleFlashColor: "#ffd27f",
+        muzzleLengthMultiplier: 1.9,
+        flashSizeMultiplier: 1.15,
+        recoilKick: 0.68,
+        recoilRecovery: 2.7,
+        turnSpeed: 6.5,
+        outline: "#56210a",
+        levelColors: [
+            "#ffcc80",
+            "#ffb74d",
+            "#ff9f43",
+            "#ff8a36",
+            "#ff7127",
+            "#ff5918",
+            "#ff4010"
+        ],
+        projectileColors: [
+            "#fff0d1",
+            "#ffe3b0",
+            "#ffd291",
+            "#ffc170",
+            "#ffb158",
+            "#ffa039",
+            "#ff9024"
+        ]
+    },
+    laser: {
+        id: "laser",
+        label: "레이저",
+        attackPattern: "laser",
+        cost: 95,
+        baseDamage: 5.5,
+        baseUpgradeCost: 80,
+        range: 200,
+        fireDelay: 0.1,
+        beamColor: "rgba(123, 234, 255, 0.9)",
+        beamGlowColor: "rgba(123, 234, 255, 0.45)",
+        beamWidth: 8,
+        sustainMultiplier: 1.15,
+        shape: "sentinel",
+        baseColor: "#0a2f3f",
+        coreColor: "#92f2ff",
+        barrelColor: "#42d8ff",
+        glowColor: "#66e7ff",
+        auraColor: "rgba(78, 228, 255, 0.32)",
+        muzzleFlashColor: "#b6f6ff",
+        muzzleLengthMultiplier: 1.85,
+        flashSizeMultiplier: 0.8,
+        recoilKick: 0.28,
+        recoilRecovery: 3.8,
+        turnSpeed: 9.5,
+        outline: "#0b3c4f",
+        levelColors: [
+            "#b1f4ff",
+            "#9be9ff",
+            "#84deff",
+            "#6dd2ff",
+            "#55c6ff",
+            "#3ab8ff",
+            "#24a7f5"
+        ],
+        projectileColors: [
+            "#e6fdff",
+            "#d0f8ff",
+            "#bbf3ff",
+            "#a4ecff",
+            "#90e5ff",
+            "#79ddff",
+            "#61d3ff"
+        ]
+    },
+    mortar: {
+        id: "mortar",
+        label: "박격포",
+        attackPattern: "mortar",
+        cost: 115,
+        baseDamage: 48,
+        baseUpgradeCost: 90,
+        range: 260,
+        fireDelay: 2.8,
+        projectileSpeed: 360,
+        projectileLife: 2.3,
+        projectileRadius: 9,
+        explosionRadius: 120,
+        explosionColor: "rgba(255, 220, 120, 0.35)",
+        explosionHaloColor: "#ffe6bd",
+        explosionStrokeColor: "rgba(255, 176, 84, 0.5)",
+        explosionLife: 0.6,
+        trailLength: 45,
+        gravity: 720,
+        mortarLift: 420,
+        detonateOnExpire: true,
+        shape: "artillery",
+        baseColor: "#36250f",
+        coreColor: "#f4d29a",
+        barrelColor: "#e88f32",
+        glowColor: "#ffc874",
+        auraColor: "rgba(255, 200, 116, 0.32)",
+        muzzleFlashColor: "#ffe1a8",
+        muzzleLengthMultiplier: 1.75,
+        flashSizeMultiplier: 1.2,
+        recoilKick: 0.72,
+        recoilRecovery: 2.5,
+        turnSpeed: 6.1,
+        outline: "#3a2b15",
+        levelColors: [
+            "#fbe0b2",
+            "#f7cb86",
+            "#f4b25a",
+            "#f19b38",
+            "#ed8320",
+            "#e26e17",
+            "#d35d12"
+        ],
+        projectileColors: [
+            "#fff2d8",
+            "#ffe4b3",
+            "#ffd28d",
+            "#ffc167",
+            "#ffad3c",
+            "#ff9a24",
+            "#ff8610"
+        ]
+    }
+};
+
+
+
+export const TOWER_ORDER = [
+    'basic',
+    'shotgun',
+    'longrange',
+    'burst',
+    'rapid',
+    'explosive',
+    'laser',
+    'mortar'
+];
+
